Fix 1.13.2 events + window title
