+++ src/mc1.13.2/java/me/dreamhopping/pml/transformers/GuiNewChatTransformer.java
@@ -16,7 +16,7 @@ public class GuiNewChatTransformer implements RuntimeTransformer {
                 list.add(new VarInsnNode(Opcodes.ALOAD, 1));
                 list.add(new MethodInsnNode(Opcodes.INVOKESTATIC, getImplementationClass("GuiNewChat"), "printChatMessage", "(Lnet/minecraft/util/text/ITextComponent;)V"));
 
-                methodNode.instructions.insertBefore(methodNode.instructions.getFirst(), list);
+                methodNode.instructions.insert(list);
             }
         }
